Extract OAuth2 superclass from TerraClient
